+++ src/Aks/Aks/ChangeLog.md
@@ -27,6 +27,7 @@
 * Added parameter `-SpotMaxPrice` for `New-AzAksNodePool`
 * Added parameter `-EnableFIPS` for `New-AzAksCluster` and `New-AzAksNodePool`
 * Added parameter `-AutoScalerProfile` for `New-AzAksCluster` and `Set-AzAksCluster`
+* Added parameter `-GpuInstanceProfile` for `New-AzAksCluster` and `New-AzAksNodePool`
 
 ## Version 5.1.0
 * Bumped API version to 2022-09-01